# backend/docs/swagger.yaml
definitions:
...
  handlers.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: This is a sample API for KU-Work
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: KU-Work API
  version: "1.0"
paths:
...
  /auth/admin/login:
    post:
      consumes:
      - application/json
      description: Authenticates an admin user with their username and password. On
        successful authentication, it returns a JWT token for session management and
        sets a refresh token in a cookie.
      parameters:
      - description: Login Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            properties:
              ' role':
                type: string
              ' userId':
                type: string
              ' username':
                type: string
              token:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: 'Unauthorized: Invalid credentials'
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Admin login
      tags:
      - Authentication
  /auth/company/login:
    post:
      consumes:
      - application/json
      description: Authenticates a company user with their username and password.
        On successful authentication, it returns a JWT token for session management
        and sets a refresh token in a cookie.
      parameters:
      - description: Login Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            properties:
              ' role':
                type: string
              ' userId':
                type: string
              ' username':
                type: string
              token:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: 'Unauthorized: Invalid credentials'
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Company login
      tags:
      - Authentication
  /auth/company/register:
    post:
      consumes:
      - multipart/form-data
      description: Handles the registration of a new company account. It takes company
        details and credentials, creates a new user and company profile, and returns
        JWT tokens upon successful registration.
      parameters:
      - description: Company's username
        in: formData
        name: username
        required: true
        type: string
      - description: Password (min 8 characters)
        in: formData
        name: password
        required: true
        type: string
      - description: Company's contact email
        in: formData
        name: email
        required: true
        type: string
      - description: Company's website URL
        in: formData
        name: website
        type: string
      - description: Company's contact phone number
        in: formData
        name: phone
        required: true
        type: string
      - description: Company's physical address
        in: formData
        name: address
        required: true
        type: string
      - description: City
        in: formData
        name: city
        required: true
        type: string
      - description: Country
        in: formData
        name: country
        required: true
        type: string
      - description: Company's profile photo
        in: formData
        name: photo
        required: true
        type: file
      - description: Company's banner image
        in: formData
        name: banner
        required: true
        type: file
      - description: About the company
        in: formData
        name: about
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Registration successful
          schema:
            properties:
              ' role':
                type: string
              ' userId':
                type: string
              ' username':
                type: string
              token:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: 'Conflict: Username already exists'
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Register a new company
      tags:
      - Authentication
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"